+++ pygnssutils/globals.py
@@ -37,3 +37,32 @@ GNSSLIST = {
     5: "QZSS",
     6: "GLONASS",
 }
+
+FIXES = {
+    "3D": 1,
+    "2D": 2,
+    "RTK FIXED": 4,
+    "RTK FLOAT": 5,
+    "RTK": 5,
+    "DR": 6,
+    "NO FIX": 0,
+}
+
+HTTPERR = [
+    "400 Bad Request",
+    "401 Unauthorized",
+    "403 Forbidden",
+    "404 Not Found",
+    "405 Method Not Allowed",
+    "406 Not Acceptable",
+]
+
+# ranges for ubxsetrate CLI
+ALLNMEA = "allnmea"
+ALLUBX = "allubx"
+MINNMEA = "minnmea"
+MINUBX = "minubx"
+ALLNMEA_CLS = [b"\xF0", b"\xF1"]
+MINMMEA_ID = [b"\xF0\x00", b"\xF0\x02", b"\xF0\x03", b"\xF0\x04", b"\xF0\x05"]
+ALLUBX_CLS = [b"\x01"]
+MINUBX_ID = [b"\x01\x07", b"\x01\x35"]